+++ test/Twilio.Test/Rest/Routes/V2/PhoneNumberResourceTest.cs
@@ -21,42 +21,6 @@ namespace Twilio.Tests.Rest.Routes.V2
     public class PhoneNumberTest : TwilioTest
     {
         [Test]
-        public void TestCreateRequest()
-        {
-            var twilioRestClient = Substitute.For<ITwilioRestClient>();
-            var request = new Request(
-                HttpMethod.Post,
-                Twilio.Rest.Domain.Routes,
-                "/v2/PhoneNumbers/phone_number",
-                ""
-            );
-            twilioRestClient.Request(request).Throws(new ApiException("Server Error, no content"));
-
-            try
-            {
-                PhoneNumberResource.Create("phone_number", client: twilioRestClient);
-                Assert.Fail("Expected TwilioException to be thrown for 500");
-            }
-            catch (ApiException) {}
-            twilioRestClient.Received().Request(request);
-        }
-
-        [Test]
-        public void TestCreateResponse()
-        {
-            var twilioRestClient = Substitute.For<ITwilioRestClient>();
-            twilioRestClient.AccountSid.Returns("ACXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX");
-            twilioRestClient.Request(Arg.Any<Request>())
-                            .Returns(new Response(
-                                         System.Net.HttpStatusCode.Created,
-                                         "{\"phone_number\": \"+18001234567\",\"url\": \"https://routes.twilio.com/v2/PhoneNumbers/+18001234567\",\"sid\": \"QQaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\",\"account_sid\": \"ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\",\"friendly_name\": \"friendly_name\",\"voice_region\": \"au1\",\"date_created\": \"2015-07-30T20:00:00Z\",\"date_updated\": \"2015-07-30T20:00:00Z\"}"
-                                     ));
-
-            var response = PhoneNumberResource.Create("phone_number", client: twilioRestClient);
-            Assert.NotNull(response);
-        }
-
-        [Test]
         public void TestUpdateRequest()
         {
             var twilioRestClient = Substitute.For<ITwilioRestClient>();
@@ -66,13 +30,11 @@ namespace Twilio.Tests.Rest.Routes.V2
                 "/v2/PhoneNumbers/phone_number",
                 ""
             );
-            request.AddPostParam("VoiceRegion", Serialize("voice_region"));
-            request.AddPostParam("FriendlyName", Serialize("friendly_name"));
             twilioRestClient.Request(request).Throws(new ApiException("Server Error, no content"));
 
             try
             {
-                PhoneNumberResource.Update("phone_number", "voice_region", "friendly_name", client: twilioRestClient);
+                PhoneNumberResource.Update("phone_number", client: twilioRestClient);
                 Assert.Fail("Expected TwilioException to be thrown for 500");
             }
             catch (ApiException) {}
@@ -90,7 +52,7 @@ namespace Twilio.Tests.Rest.Routes.V2
                                          "{\"phone_number\": \"+18001234567\",\"url\": \"https://routes.twilio.com/v2/PhoneNumbers/+18001234567\",\"sid\": \"QQaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\",\"account_sid\": \"ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\",\"friendly_name\": \"friendly_name\",\"voice_region\": \"au1\",\"date_created\": \"2015-07-30T20:00:00Z\",\"date_updated\": \"2015-07-30T20:00:00Z\"}"
                                      ));
 
-            var response = PhoneNumberResource.Update("phone_number", "voice_region", "friendly_name", client: twilioRestClient);
+            var response = PhoneNumberResource.Update("phone_number", client: twilioRestClient);
             Assert.NotNull(response);
         }